refactor(app): extract server dialog owners
